+++ jlegmed-test/src/main/java/io/jexxa/jlegmedtest/architecture/PackageName.java
@@ -9,10 +9,15 @@ final class PackageName
 {
     public static final String DOMAIN_EVENT = "..domainevent..";
     public static final String TELEMETRY_DATA = "..dto.telemetrydata..";
+    @Deprecated
     public static final String PUBLISHED_MESSAGE = "..dto.publishedmessage..";
     public static final String CONSUMED_MESSAGE = "..dto.consumedmessage..";
+    public static final String PRODUCED_MESSAGE = "..dto.producedmessage..";
+    @Deprecated
     public static final String FLOW_DATA = "..dto.flowdata..";
+    public static final String FLOW_CONTRACT = "..flowgraph.contract..";
     public static final String FLOW_ERROR = "..dto.flowerror..";
+    public static final String FLOW_CONTRACT_ERROR = "..flowgraph.error..";
     public static final String PLUGINS = "..plugins..";
 
     private PackageName()