+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## \[5.2.x] - 2026-mm-dd
+
+### Added 
+- Added annotations 
+  - `@EventPayload`: Event Payload
+  - `@FlowContract`: Replaces `@FlowData` to represend data flow inside the application  
+  - `@ProducedMessage`: Replaces `@PublishedMessage`
+  - `@MessagePayload`: For consumed and produced messages
+
+### Changed
+- Annotations declared as deprecated
+  - `@FlowData` (replaced by `@FlowContract`)
+  - `@PublishedMessage` (replaced by `@ProducedMessage`)
+
 ## \[5.1.5] - 2026-04-03
 ### Fixed
 - Updated dependencies